requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.3
typer
rich
ollama
pytest>=7.4.0
pytest-cov>=4.1.0
httpx
